solver: split exec and source to solver/llbop pkg
