+++ solver/llbop_build.go
